backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
scikit-learn==1.5.0
numpy==1.26.4
python-dotenv==1.0.1
